requirements.txt
Django==3.1.2
djangorestframework==3.12.1
djoser==2.1.0
psycopg2==2.8.6
pillow==8.0.1
drf-yasg==1.20.0
